AvahiCode + Bonjour: use correct datatype for pos
ssize_t is not available on windows (at least not with msvc2019). Use
int as QByteArray::indexOf is defined with int return type.

+++ avahicore.cpp
@@ -183,7 +183,7 @@ public:
 				while (txt)	// get txt records
 				{
 					QByteArray avahiText((const char *)txt->text, txt->size);
-					const ssize_t pos = avahiText.indexOf('=');
+					const int pos = avahiText.indexOf('=');
 					if (pos < 0)
 						zcs->m_txt[avahiText] = "";
 					else
